# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WaStatus {
  connected: boolean;
  /** @nullable */
  qr?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  pairingCode?: string | null;
}

export interface DisconnectResult {
  success: boolean;
  message: string;
}

export interface PairingCodeInput {
  /** Phone number in international format e.g. 923001234567 */
  phone: string;
}

export interface PairingCodeResult {
  success: boolean;
  /** @nullable */
  code?: string | null;
  message: string;
}

export interface NumberInput {
  number: string;
}

export type VerifyResultStatus = typeof VerifyResultStatus[keyof typeof VerifyResultStatus];


export const VerifyResultStatus = {
  active: 'active',
  deleted: 'deleted',
  error: 'error',
} as const;

export interface VerifyResult {
  status: VerifyResultStatus;
  reason: string;
  number: string;
}

export interface StartScanInput {
  numbers: string[];
  workspaceId?: string;
  /** Delay in ms between each check (default 400) */
  delay?: number;
}

export interface StartScanResult {
  jobId: string;
  total: number;
}

export type ScanResultStatus = typeof ScanResultStatus[keyof typeof ScanResultStatus];


export const ScanResultStatus = {
  active: 'active',
  deleted: 'deleted',
  error: 'error',
} as const;

export interface ScanResult {
  number: string;
  status: ScanResultStatus;
  reason: string;
  timestamp: string;
}

export type ScanJobStatus = typeof ScanJobStatus[keyof typeof ScanJobStatus];


export const ScanJobStatus = {
  running: 'running',
  stopped: 'stopped',
  done: 'done',
} as const;

export interface ScanJob {
  id: string;
  workspaceId: string;
  numbers: string[];
  currentIndex: number;
  results: ScanResult[];
  status: ScanJobStatus;
  delay: number;
  startedAt: string;
  /** @nullable */
  stoppedAt: string | null;
}

export interface BotSetupInput {
  token: string;
}

export interface BotSetupResult {
  success: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface BotStatus {
  connected: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface SuccessResult {
  success: boolean;
}

export interface ErrorResult {
  error: string;
}