# lib/api-zod/src/generated/types/botSetupResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotSetupResult {
  success: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  error?: string | null;
}