# lib/api-zod/src/generated/types/scanJob.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanJobStatus } from './scanJobStatus';
import type { ScanResult } from './scanResult';

export interface ScanJob {
  id: string;
  workspaceId: string;
  numbers: string[];
  currentIndex: number;
  results: ScanResult[];
  status: ScanJobStatus;
  delay: number;
  startedAt: string;
  /** @nullable */
  stoppedAt: string | null;
}