# lib/api-zod/src/generated/types/startScanInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StartScanInput {
  numbers: string[];
  workspaceId?: string;
  /** Delay in ms between each check (default 400) */
  delay?: number;
}